Media View: Use usuiat/Zoomable for zooming gestures
Before I still used the same library, but it lacked some features that I needed so I've imported the library in the app project to modify to match the app's needs. Now I can track the library itself from the owner as the needed feature is now available along with some bug fixes!

Signed-off-by: IacobIonut01 <paulionut2003@gmail.com>