Gallery: Add open with option for videos
Partially fullfills the issue #11. By default the user will be able to scroll through all of the media, and for any video he wants to open externally, can be done so using the "Open with" button (same as "Use as" for images). Video autoplay is not enabled so the user can specifically open any video he wants externally without having to first pause the current playback

Signed-off-by: IacobIonut01 <paulionut2003@gmail.com>